Add translation of Case statements
Note that to solve this where:

filter' p Nil = Nil
filter' p (x `Cons` xs) = case p x of
                     True  -> x `Cons` rest
                     False -> rest
  where rest = filter' p xs

It was added that FV also looks into indirections in scope, that is,
binds in the monad, and checks for FV there, so the floated case
will have p and xs as arguments. Futhermore, when rest is compiled
it is added to the scope, and rightly so, since it is in scope, so
the floated case does not need to get a reference (with a name clash)
to this function.